requirements.txt
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
pydantic
email-validator
itsdangerous